+++ source/dotnet/Measurements.Abstractions/Api/Models/MeasurementAggregationByDateDto.cs
@@ -7,6 +7,6 @@
 /// <param name="Quantity">Quantity of aggregated measurement.</param>
 /// <param name="Quality">Value of lowest quality among measurements in aggregation.</param>
 /// <param name="Unit">Unit of aggregated measurement.</param>
-/// <param name="ContainsMissingValues">Indicates whether aggregate contains missing values during day of aggregation.</param>
+/// <param name="IsMissingValues">Indicates whether aggregate contains missing values during day of aggregation.</param>
 /// <param name="ContainsUpdatedValues">Indicates whether any measurements in aggregate has been updated since creation.</param>
-public sealed record MeasurementAggregationByDateDto(DateOnly Date, decimal? Quantity, Quality Quality, Unit Unit, bool ContainsMissingValues, bool ContainsUpdatedValues);
+public sealed record MeasurementAggregationByDateDto(DateOnly Date, decimal? Quantity, Quality Quality, Unit Unit, bool IsMissingValues, bool ContainsUpdatedValues);