+++ source/dotnet/Measurements.Application/Handlers/IMeasurementsHandler.cs
@@ -18,7 +18,8 @@ public interface IMeasurementsHandler
     /// Get measurements aggregated by date matching request attributes.
     /// </summary>
     /// <param name="getAggregatedByDateRequest"></param>
-    Task<DeprecatedMeasurementsAggregatedByDateResponse> DeprecatedGetAggregatedByDateAsync(GetAggregatedByDateRequest getAggregatedByDateRequest);
+    [Obsolete("GetAggregatedByDateAsyncV3 is deprecated. Use GetAggregatedByDateAsync instead.")]
+    Task<MeasurementsAggregatedByDateResponseV3> GetAggregatedByDateAsyncV3(GetAggregatedByDateRequest getAggregatedByDateRequest);
 
     /// <summary>
     /// Get measurements aggregated by date matching request attributes.