+++ source/dotnet/Measurements.Infrastructure/Handlers/MeasurementsHandler.cs
@@ -31,14 +31,15 @@ public class MeasurementsHandler(IMeasurementsRepository measurementsRepository)
         return MeasurementsAggregatedByDateResponse.Create(aggregatedMeasurements);
     }
 
-    public async Task<DeprecatedMeasurementsAggregatedByDateResponse> DeprecatedGetAggregatedByDateAsync(GetAggregatedByDateRequest request)
+    [Obsolete("GetAggregatedByDateAsyncV3 is deprecated. Use GetAggregatedByDateAsync instead.")]
+    public async Task<MeasurementsAggregatedByDateResponseV3> GetAggregatedByDateAsyncV3(GetAggregatedByDateRequest request)
     {
         var yearMonth = new YearMonth(request.Year, request.Month);
         var aggregatedMeasurements = await measurementsRepository
             .GetAggregatedByDateAsync(request.MeteringPointId, yearMonth)
             .ToListAsync() ?? throw new MeasurementsNotFoundException();
 
-        return DeprecatedMeasurementsAggregatedByDateResponse.Create(aggregatedMeasurements);
+        return MeasurementsAggregatedByDateResponseV3.Create(aggregatedMeasurements);
     }
 
     public async Task<MeasurementsAggregatedByMonthResponse> GetAggregatedByMonthAsync(GetAggregatedByMonthRequest request)